권한추가 및 로그인 로그아웃

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -12,6 +12,8 @@ import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 
@@ -20,12 +22,26 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession session, int id, String title, String body) {
+		
+		boolean isLogined = false;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인 후 이용바랍니다.");
+		}
 
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
+		}
+		
+		if (!session.getAttribute("loginedMemberId").equals(article.getWriter())) {
+			return ResultData.from("F-2", "수정 권한이 없습니다.");
 		}
 
 		articleService.modifyArticle(id, title, body);
@@ -50,12 +66,27 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
+		
+		boolean isLogined = false;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인 후 이용바랍니다.");
+		}
+		
 
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
+		}
+		
+		if (!session.getAttribute("loginedMemberId").equals(article.getWriter())) {
+			return ResultData.from("F-2", "삭제 권한이 없습니다.");
 		}
 
 		articleService.deleteArticle(id);
@@ -65,7 +96,18 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData doWrite(String title, String body) {
+	public ResultData doWrite(HttpSession session, String title, String body) {
+		
+		boolean isLogined = false;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인 후 이용바랍니다.");
+		}
+
 
 		if (Ut.isEmptyOrNull(title)) {
 			return ResultData.from("F-1", "제목을 입력하세요");
